LayoutTests/fast/css-grid-layout/grid-disable-expected.txt

Verifies that CSS Grid Layout does not work if runtime feature is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('grid').style.display is ''
PASS document.getElementById('inlinegrid').style.display is ''
PASS successfullyParsed is true

TEST COMPLETE
